cluster: extract incremental sync sweep into sync_conn_sweep.go
Second pure code-motion increment of the #5661 sync_conn.go split. Moves
the periodic incremental session-sync sweep and its zone-gating helpers
into a cohesive sibling file.

Moved verbatim (no rename, no logic/timing change):
  StartSyncSweep, sweepIntervals, sweepIntervalsForDataPlane,
  ShouldSyncZone, syncSweep.

Same package (pkg/cluster); imports recomputed per file. Behavior is
byte-identical.

go build ./... passes.
